+++ database/__init__.py
@@ -1,3 +1,3 @@
-from .postgres import Database
+from .database import Database
 
 __all__ = ('Database',)