+++ settings/browser_set.py
@@ -16,6 +16,7 @@ browser_launch_options = {
     'firefox_user_prefs': {
         # Отключить детекцию webdriver
         'dom.webdriver.enabled': False,
+
         # Отключить телеметрию
         'toolkit.telemetry.enabled': False,
         'toolkit.telemetry.unified': False,